restruct modules and cleaned
